requirements.txt
tensorflow==2.13.0
keras==2.13.1
opencv-python==4.8.1
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
scikit-image==0.21.0
imbalanced-learn==0.11.0
tqdm==4.65.0
pyyaml==6.0.1
albumentations==1.3.1
jupyter
ipykernel
